[libs][core] Cleaning up tree utils
